updated client to request correct parameters

+++ Project_4/client.sh
@@ -7,8 +7,13 @@ function ctrl_c() {
 	exit 0
 }
 
+if [ "$#" -ne 2 ]; then
+    echo "Usage is: $0 <address> <port>"
+    exit 1
+fi
+
 mkfifo upload
-echo 'Use cat "<filename> >>bin" to send a file'
+echo 'Use "cat <filename> >>upload" to send a file'
 nc $1 $2 <upload &
 cat >>upload
 rm upload
